тачки ищем

+++ Models/CarEngine.cs
@@ -6,6 +6,6 @@
 
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Capacity { get; set; }
+        public float Capacity { get; set; }
     }
 }
